7. Average score

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -7,7 +7,7 @@ DROP PROCEDURE IF EXISTS ComputeAverageScoreForUser;
 DELIMITER |
 CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-    DECLARE avg_score INT;
+    DECLARE avg_score  DECIMAL(10,2);
     SELECT (SUM(score) / COUNT(project_id)) INTO avg_score FROM corrections
         WHERE corrections.user_id = user_id;
     UPDATE users SET average_score = avg_score WHERE id = user_id;
